+++ tests/specs/typeAlias/literal/boolean.d.ts
@@ -1,0 +1,3 @@
+type Test2 = true;
+
+type Test3 = false;